rearrange utility functions
